+++ CHANGES.md
@@ -24,6 +24,7 @@ Changes in 13.0
 - Add RosarioSIS\Functions\Hacking class in Hacking.php
 - Use RosarioSIS\Functions\Hacking class, program wide
 - Remove deprecated `_printPageHead()` function in PasswordReset.php
+- Deprecate Linkify() function in Linkify.fnc.php
 
 Changes in 12.9.2
 -----------------